server/requirements.txt
Django~=4.2.2
djangorestframework~=3.14.0
psycopg2~=2.9.6
drf-spectacular~=0.26.3
web3~=6.5.0

#for local environment
# pytest-django~=4.5.2
# eth-tester~=0.9.0b1
